backend/userside/requirements.txt
asgiref==3.7.2
dj-rest-auth-social==2.2.4
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
gunicorn==21.2.0